+++ cli/src/commands/mod.rs
@@ -2,6 +2,7 @@ use std::process;
 
 pub mod auth;
 pub mod extensions;
+pub mod find_lockable_files;
 pub mod group;
 pub mod init;
 pub mod jobs;